Implement initial module and CLI (#1)
The `run` method runs tests with `prove`. It takes a config object and returns
an `EventEmitter`. Currently, it is possible to configure the level of
concurrency and the command to run tests with. It's also possible to specify JS
modules to run before and after the tests.

The CLI wraps `run` and is installed as `proving-ground`.

An example (runnable via the CLI) can be found in examples/leadfoot/.